+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v0.2.0 (in development)
   objects belonging to the month
 - Gave `Date` `pred()` and `succ()` methods for getting the previous & next
   dates
+- Added a `Weekday` enum and a `Date::weekday()` method
 
 v0.1.0 (2023-05-02)
 -------------------